# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Auriga backend API — waitlist, device telemetry, hazard events, field tests
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface WaitlistInput {
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  source?: string | null;
}

export interface WaitlistEntry {
  id: number;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  source?: string | null;
  createdAt: string;
}

export interface CountStat {
  label: string;
  count: number;
}

export interface WaitlistStats {
  total: number;
  byCountry: CountStat[];
  bySource: CountStat[];
}

export interface DeviceInput {
  deviceName: string;
  androidVersion: string;
  ramGb: number;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  testerName?: string | null;
}

export interface Device {
  id: number;
  deviceName: string;
  androidVersion: string;
  ramGb: number;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  testerName?: string | null;
  status: string;
  /** @nullable */
  lastSeenAt?: string | null;
  createdAt: string;
}

export interface HazardInput {
  deviceId: number;
  hazardType: string;
  confidence: number;
  /** @nullable */
  distanceMeters?: number | null;
  /** @nullable */
  bearingDegrees?: number | null;
  /** @nullable */
  audioAlertFired?: boolean | null;
  /** @nullable */
  hapticAlertFired?: boolean | null;
  /** @nullable */
  notes?: string | null;
}

export interface HazardEvent {
  id: number;
  deviceId: number;
  hazardType: string;
  confidence: number;
  /** @nullable */
  distanceMeters?: number | null;
  /** @nullable */
  bearingDegrees?: number | null;
  /** @nullable */
  audioAlertFired?: boolean | null;
  /** @nullable */
  hapticAlertFired?: boolean | null;
  /** @nullable */
  notes?: string | null;
  detectedAt: string;
}

export interface HazardSummary {
  totalEvents: number;
  avgConfidence: number;
  byType: CountStat[];
  recentEvents: HazardEvent[];
}

export interface FieldSessionInput {
  deviceId: number;
  locationName: string;
  durationMinutes: number;
  /** @nullable */
  hazardsDetected?: number | null;
  /** @nullable */
  modelLoaded?: boolean | null;
  /** @nullable */
  cameraWorked?: boolean | null;
  /** @nullable */
  audioWorked?: boolean | null;
  /** @nullable */
  notes?: string | null;
}

export interface FieldSession {
  id: number;
  deviceId: number;
  locationName: string;
  durationMinutes: number;
  /** @nullable */
  hazardsDetected?: number | null;
  /** @nullable */
  modelLoaded?: boolean | null;
  /** @nullable */
  cameraWorked?: boolean | null;
  /** @nullable */
  audioWorked?: boolean | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  waitlistTotal: number;
  devicesTotal: number;
  hazardsTotal: number;
  sessionsTotal: number;
  recentHazards: HazardEvent[];
  topHazardTypes: CountStat[];
}

export type ListHazardsParams = {
/**
 * @nullable
 */
deviceId?: number | null;
/**
 * @nullable
 */
limit?: number | null;
};