# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Auriga backend API — waitlist, device telemetry, hazard events, field tests
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /devices:
    get:
      operationId: listDevices
      tags: [devices]
      summary: List all registered field test devices
      responses:
        "200":
          description: Device list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      operationId: registerDevice
      tags: [devices]
      summary: Register a new field test device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceInput"
      responses:
        "201":
          description: Device registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{deviceId}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get device details
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /hazards:
    get:
      operationId: listHazards
      tags: [hazards]
      summary: List hazard detection events
      parameters:
        - name: deviceId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Hazard events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HazardEvent"
    post:
      operationId: reportHazard
      tags: [hazards]
      summary: Report a hazard detection event from device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HazardInput"
      responses:
        "201":
          description: Hazard recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HazardEvent"
...
components:
  schemas:
...
    DeviceInput:
      type: object
      required: [deviceName, androidVersion, ramGb]
      properties:
        deviceName:
          type: string
        androidVersion:
          type: string
        ramGb:
          type: number
        location:
          type: ["string", "null"]
        testerName:
          type: ["string", "null"]
    Device:
      type: object
      required: [id, deviceName, androidVersion, ramGb, status, createdAt]
      properties:
        id:
          type: integer
        deviceName:
          type: string
        androidVersion:
          type: string
        ramGb:
          type: number
        location:
          type: ["string", "null"]
        testerName:
          type: ["string", "null"]
        status:
          type: string
        lastSeenAt:
          type: ["string", "null"]
        createdAt:
          type: string
    HazardInput:
      type: object
      required: [deviceId, hazardType, confidence]
      properties:
        deviceId:
          type: integer
        hazardType:
          type: string
        confidence:
          type: number
        distanceMeters:
          type: ["number", "null"]
        bearingDegrees:
          type: ["number", "null"]
        audioAlertFired:
          type: ["boolean", "null"]
        hapticAlertFired:
          type: ["boolean", "null"]
        notes:
          type: ["string", "null"]
    HazardEvent:
      type: object
      required: [id, deviceId, hazardType, confidence, detectedAt]
      properties:
        id:
          type: integer
        deviceId:
          type: integer
        hazardType:
          type: string
        confidence:
          type: number
        distanceMeters:
          type: ["number", "null"]
        bearingDegrees:
          type: ["number", "null"]
        audioAlertFired:
          type: ["boolean", "null"]
        hapticAlertFired:
          type: ["boolean", "null"]
        notes:
          type: ["string", "null"]
        detectedAt:
          type: string
...