# lib/api-zod/src/generated/types/deviceInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Auriga backend API — waitlist, device telemetry, hazard events, field tests
 * OpenAPI spec version: 0.1.0
 */

export interface DeviceInput {
  deviceName: string;
  androidVersion: string;
  ramGb: number;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  testerName?: string | null;
}