# lib/api-zod/src/generated/types/hazardEvent.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Auriga backend API — waitlist, device telemetry, hazard events, field tests
 * OpenAPI spec version: 0.1.0
 */

export interface HazardEvent {
  id: number;
  deviceId: number;
  hazardType: string;
  confidence: number;
  /** @nullable */
  distanceMeters?: number | null;
  /** @nullable */
  bearingDegrees?: number | null;
  /** @nullable */
  audioAlertFired?: boolean | null;
  /** @nullable */
  hapticAlertFired?: boolean | null;
  /** @nullable */
  notes?: string | null;
  detectedAt: string;
}